Fix some bugs

+++ common/register.sh
@@ -5,4 +5,8 @@ get_id(){
     xxd -s 0x60 -l 8 -g 8 -u -ps $1
 }
 chmod 600 $DEVICE_INFO_IMG
-mv -f $DEVICE_INFO_IMG $REGISTRATIONDIR/$(get_id $DEVICE_INFO_IMG)
+if [ -s $DEVICE_INFO_IMG ]; then
+    mv -f $DEVICE_INFO_IMG $REGISTRATIONDIR/$(get_id $DEVICE_INFO_IMG)
+else
+    return 1
+fi
